refactor: make embed toolbar as a widget
